requirement.txt
# Core
Django==2.0.7

# Rest apis
djangorestframework==3.8.2
requests==2.19.1

# DB
psycopg2==2.7.4
djangorestframework-datatables==0.5.0
